+++ RELEASE_NOTES.md
@@ -44,7 +44,7 @@
 
 ## v2.16.0 (2024-04)
 
-* Handle deprecation of the `airport` command starting at Mac OS 14.4.
+* Handle deprecation of the `airport` command starting at macOS 14.4.
 * Add hotspot_login_required functionality.
 * Change 'port' to 'interface' in some names.
 * Add to external resources: captive.apple.com, librespeed.org